refactor: clean api endpoints
