+++ src/pages/api/check.ts
@@ -1,9 +1,11 @@
 import { NextApiRequest, NextApiResponse } from 'next';
-import { checkUserExists } from '../../utils/fetch';
+import checkUserExists from '../../utils/api/check-user-exists';
 
 const check = async (req: NextApiRequest, res: NextApiResponse) => {
   const { user } = req.query;
+
   const data = await checkUserExists(String(user));
+
   res.setHeader('Cache-Control', 's-maxage=3600, stale-while-revalidate');
   res.json(data);
 };
