+++ src/pages/api/contributions.ts
@@ -1,9 +1,11 @@
 import { NextApiRequest, NextApiResponse } from 'next';
-import { fetchData } from '../../utils/fetch';
+import fetchGameData from '../../utils/api/fetch-game-data';
 
 const contributions = async (req: NextApiRequest, res: NextApiResponse) => {
   const { user, year } = req.query;
-  const data = await fetchData(String(user), Number(year));
+
+  const data = await fetchGameData(String(user), Number(year));
+
   res.setHeader('Cache-Control', 's-maxage=3600, stale-while-revalidate');
   res.json(data);
 };
